+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
   * Added missing run-time dependencies to `franka_description` and `franka_control`
   * Added epsilon to grasp action
+  * Added `m_ee`, `F_x_Cee`, `I_ee`, `m_total`, `F_x_Ctotal` and `I_total`
+    to the robot state
   * Use O_T_EE_d in examples
 
 ## 0.1.2 - 2017-10-10